Skip tests needing NCBI BLAST+

+++ tests/test_assess.sh
@@ -43,29 +43,34 @@ set +o pipefail
 thapbi_pict assess -d $DB -m identity --input tests/assess/*.identity.tsv tests/assess/*.known.tsv 2>&1 | grep "WARNING: 1 expected species were not a possible prediction: Phytophthora fallax"
 set -o pipefail
 
-# Originally these were created in tests/test_classify.sh
-if [ ! -f $TMP/DNAMIX_S95_L001.blast.tsv ]; then
-    mkdir -p $TMP/thapbi_blast/
-    thapbi_pict classify -m blast -i tests/prepare-reads/DNAMIX_S95_L001.fasta -o $TMP/thapbi_blast/
+
+if ! [ -x "$(command -v blastn)" ]; then
+    echo 'WARNING: NCBI BLAST+ not installed, skipping some tests'
+else
+    # Originally these were created in tests/test_classify.sh
+    if [ ! -f $TMP/DNAMIX_S95_L001.blast.tsv ]; then
+        mkdir -p $TMP/thapbi_blast/
+        thapbi_pict classify -m blast -i tests/prepare-reads/DNAMIX_S95_L001.fasta -o $TMP/thapbi_blast/
+    fi
+    if [ ! -f $TMP/DNAMIX_S95_L001.identity.tsv ]; then
+        thapbi_pict classify -m identity -i tests/prepare-reads/DNAMIX_S95_L001.fasta -o $TMP/
+    fi
+
+    rm -rf $TMP/assess_blast_vs_identity.tsv
+    rm -rf $TMP/confusion_blast_vs_identity.tsv
+
+    echo "Testing blast vs identity"
+    # Note using the default DB here, as would be used in test_classify.sh to generate inputs
+    # Don't have a gold standard known truth to test this against, so test blast vs identity
+    thapbi_pict assess -i $TMP/thapbi_blast/DNAMIX_S95_L001.blast.tsv $TMP/DNAMIX_S95_L001.identity.tsv -m blast --known identity -o $TMP/assess_blast_vs_identity.tsv -c $TMP/confusion_blast_vs_identity.tsv
+
+    # Check assessment output to stdout works (default):
+    thapbi_pict assess -i $TMP/thapbi_blast/DNAMIX_S95_L001.blast.tsv $TMP/DNAMIX_S95_L001.identity.tsv -m blast --known identity > $TMP/stdout.txt
+    diff $TMP/stdout.txt $TMP/assess_blast_vs_identity.tsv
+
+    # Check confusion matrix output to stdout works, also give one input as a directory name
+    thapbi_pict assess -i $TMP/thapbi_blast $TMP/DNAMIX_S95_L001.identity.tsv --method blast --known identity -o $TMP/assess_blast_vs_identity.tsv -c - > $TMP/stdout.txt
+    diff $TMP/stdout.txt $TMP/confusion_blast_vs_identity.tsv
 fi
-if [ ! -f $TMP/DNAMIX_S95_L001.identity.tsv ]; then
-    thapbi_pict classify -m identity -i tests/prepare-reads/DNAMIX_S95_L001.fasta -o $TMP/
-fi
-
-rm -rf $TMP/assess_blast_vs_identity.tsv
-rm -rf $TMP/confusion_blast_vs_identity.tsv
-
-echo "Testing blast vs identity"
-# Note using the default DB here, as would be used in test_classify.sh to generate inputs
-# Don't have a gold standard known truth to test this against, so test blast vs identity
-thapbi_pict assess -i $TMP/thapbi_blast/DNAMIX_S95_L001.blast.tsv $TMP/DNAMIX_S95_L001.identity.tsv -m blast --known identity -o $TMP/assess_blast_vs_identity.tsv -c $TMP/confusion_blast_vs_identity.tsv
-
-# Check assessment output to stdout works (default):
-thapbi_pict assess -i $TMP/thapbi_blast/DNAMIX_S95_L001.blast.tsv $TMP/DNAMIX_S95_L001.identity.tsv -m blast --known identity > $TMP/stdout.txt
-diff $TMP/stdout.txt $TMP/assess_blast_vs_identity.tsv
-
-# Check confusion matrix output to stdout works, also give one input as a directory name
-thapbi_pict assess -i $TMP/thapbi_blast $TMP/DNAMIX_S95_L001.identity.tsv --method blast --known identity -o $TMP/assess_blast_vs_identity.tsv -c - > $TMP/stdout.txt
-diff $TMP/stdout.txt $TMP/confusion_blast_vs_identity.tsv
 
 echo "$0 - test_assess.sh passed"
